Fixes 'Please Wait' popup when for no PT images
Fixes an annoying issue where trying to restore a filesystem-directly-on-disk
image has a 'Please Wait' popup (for unmounting the LVMs) which never
disappears.

It looks like the self.please_wait_popup variable is having its reference
clobbered due to calling the function twice, and after unmounting the LVM both
callback functions cleanup (but due the clobbering only one please wait popup
is destroyed).

A simple workaround is to cleanup the popup before creating another. This is
not an ideal fix, but is sufficient for now.

https://github.com/rescuezilla/rescuezilla/issues/43