Fixes progress bar, temp_dir, logger init order
Ensures temporary directory variable and logger is constructed at the correct
times. BackupManager and RestoreManager needed a logger available as it was
used during eg, updating progress bar. The simplest fix is for those functions
to check if the logger is 'None' before using it. Then always reset the
progress bar before backup/restore/clone starts.