Restores MBR gap using MBR for some image formats
Handles the case where the post-MBR gap is part of the MBR file itself. A
design decision universally taken outside of the Clonezilla image format.

Also takes care to ensure the post-MBR gap is not overwritten for GPT disks, as
this can cause an inconsistent state after the post-MBR gap changes.