Generates language pack package lists
Refactors the list of language pack Ubuntu packages to install for applications
like the Firefox web browser by simply listing the language codes as an array
and generating the list of packages to install.

Previously the package lists were being listed explicitly, but as the number of
supported languages grows this has become unwieldy and would eventually become
error prone.

The list of language codes list is still duplicated between [1], [2] and [3],
but this is easily manageable and not a problem.

[1] src/livecd/image/boot/grub/grub.cfg

[2] chroot.steps.part.1.sh

[3] README.md