Add support for images created by Apart GTK
Adds the ability to restore and explore images created by the Apart GTK
partclone frontend.

Testing consisted of a small number of images created by Apart GTK. Further
testing will be conducted with a larger corpus and any bugs can be fixed as
commits on top of this one.

[1] https://github.com/rescuezilla/rescuezilla/issues/230

Fixes #230