Fixes 'already scanned' image logic
Fixes and simplifies the logic around images which have been already scanned by
using dirname (the name of the folder) instead of basename (the filename) and
considering the scanned folder set as an ignore list, which is only used to
prevent double scanning of images.

Also removes and simplifies the is_subfolder/filename parameters in the
scan_file() function signature, because these variables are no longer needed.

This may address issues where users see missing images during a scan [1]

[1] https://github.com/rescuezilla/rescuezilla/issues/251