Displays error on duplicate backup image name
Displays an error when the user specifies an image name that already exists.

Previously this would only be detected during the backup process itself, and
the user would need to run through the entire backup wizard again.

Note: the error message is tentative and may be modified before submitting to
translators on Weblate.

https://github.com/rescuezilla/rescuezilla/issues/238

Fixes #238