Ensures failed LVM shutdown prints error correctly
Ensures that on a failed LVM shutdown there are some details available for the
user to read.

By using the output of the Utility.run() command there might be a bit of
duplication around the introduction of the error message:

    Failed to shutdown Logical Volume Manager (LVM) Volume Group (VG): vgtest2

    Failed to run command: vgchange --activate n --ignorelockingfailure --partial vgtest2

    <explanation of error>

But this 'failed to' duplication far more preferable to the exception shown in
issue #250.

Fixes #250