Removes RedoBackupLegacyImage unit test stub
Removes the unit test stub used during early development to support the Redo
Backup / Rescuezilla v1.0.5-v1.0.6.1 images. The "unit test" is not actually
asserting anything, so it just adds noise. It also relies on a test image
that's not part of the git repository.

In future, a unit test covering such images containing actual asserts may be
implemented.

[1] https://github.com/rescuezilla/rescuezilla/issues/151