Adds installers for integration test TCP server
Adds a shell script and batch script to install the Rescuezilla Integration
Test Server as a service (systemd service for Linux, and NSSM for Windows)

This TCP server allows the Rescuezilla Integration Test script to verify that
the target machine successfully boots into the operating system (which means
the restore operation was successful).

It's expected that the install scripts will be run in a VirtualBox environment
during the *manual* preparation steps involved creating and integrating test
images into the Rescuezilla's Integration Test Suite. Once integrated, the test
suite allows automated verification.

https://github.com/rescuezilla/rescuezilla/issues/151