Adds listip integration test script
Adds command to list the static DHCP reservation of the virtual machine
configured in Rescuezilla's Integration Test Suite.

This makes it easier for a scripts that don't have access to the constants of
the Python script (such as a test written in Bash) to be able to connect to the
VMs.

https://github.com/rescuezilla/rescuezilla/issues/151