Adds Clonezilla end-to-end integration tests
Adds integration test bash scripts which leverages the integration_test.py
script to launch the VMs, and then connect to the VM through SSH and run the
Clonezilla command-line interface to backup and restore images (zeroing the
drive in between).

The tests are done against four environments: CentOS, Ubuntu/Windows dual boot,
Windows BIOS and Windows EFI.

This is a potentially dangerous operation, so care has been taken to ensure the
operation is taking place within a VirtualBox environment.

Rescuezilla doesn't yet have a command-line [2], so the tests currently only
use Clonezilla and will be expanded in a future commit.

https://github.com/rescuezilla/rescuezilla/issues/151

https://github.com/rescuezilla/rescuezilla/issues/258