Completes initial image verify implementation
Completes the initial implementation of the Verify [1] feature. The wizard
skeleton was originally added in [2] but disabled before Rescuezilla v2.2
release, while support for other image formats was being implemented.

Further work is needed to test the implementation over a wide variety of image
files, and to add the translation skeleton so the Verify wizard can be
localized into many different languages.

[1] https://github.com/rescuezilla/rescuezilla/issues/30

[2] https://github.com/rescuezilla/rescuezilla/commit/2cda4a9cddb8d44daf45be24a64585a05328e0cd