Stores estimated bytes for Foxclone partitions
Stores the partition's estimated bytes (from the partition table) for Foxclone
images, as the lack of this key in the dictionary was causing errors.

This is a quick hotfix, to continue developing other features. This change may
be built upon, further tested and polished in subsequent commits based on
testing with a larger variety of Foxclone images.