Displays serial number of drive as column
Displays the serial number of the drive as a new columnm (next to the existing
model number) on each wizard page that refers to a drive (rather than a
partition).

This helps delineate the situation where multiple identical drives are being
cloned.

This doesn't directly address [1] (wanting the model to be associated with the
partition, eg when selecting an image to restore), but displaying the drive
serial number later in the wizard provides a simple avenue for users to quickly
delineate the situation.

Fixes #227

[1] https://github.com/rescuezilla/rescuezilla/issues/240