Splits combined label: verify another/new session
Splits up a mistakenly combined label for the new verify feature that asks the
user to click back to verify another image, and click next to start new
session.

The reason for this separation is it simplifies translation, as the "start new
session" text is already translated and therefore can be automatically used.

https://github.com/rescuezilla/rescuezilla/issues/30