Adds 'rescue' option for Backup/Restore/Clone
Adds a 'rescue' checkbox option to the confirmation page for
Backup/Restore/Clone operations, to allow the user to enable the following
partclone options

* "--force" (ignore filesystem inconsistencies)

* "--rescue" (ignore drive read errors)

This addresses a long-running complaint where Rescuezilla had poor handling of
disks with such issues.

The checkbox on the confirmation page is not ideal, but it's sufficient for
now.

Fixes #237