Ensures .travis.yml trusts rescuezilla git
Runs the command to 'trust' the Rescuezilla git repository.

The .travis.yml build bot script is referenced in the build documentation for
building using Docker. The project does not currently build on the version of
git that's currently in the Dockerfile, as git's default security restrictions
appear to have been tightened.