Add NFS version selection: unspecified, v3, v4
Adds the ability to specify a version argument with the "mount.nfs" call, by
leveraging the same combobox as the SMB/CIFS case but populating it with
NFS-specific strings

The previous behavior is now called 'unspecified' and continues to be the
default behavior.

From `man 5 nfs`

nfsvers=n      The NFS protocol version number used to contact the server's NFS
service.  If the server does not support the  requested  version,  the  mount
request fails.  If this option is not specified, the client tries version 4.2
first, then negotiates down until it finds a version supported by the server.

vers=n         This option is an alternative to the nfsvers option.  It is
included for compatibility with other operating systems

Translation strings left for subsequent commit for reviewability.

Adapted from changes proposed by OAM7575 in PR #408.

Issue: #408, #354

Co-authored-by: OAM7575 <oam7575@gmail.com>

Co-authored-by: Shasheen Ediriweera <shasheene@gmail.com>