Init SMB/CIFS vers combobox on selection/on-change
Moves initialization of the dropdown for the SMB/CIFS network folder mounting
to after the SMB/CIFS key has been selected (on-change), rather than upon
startup.

This provides a tidy way for the combobox to be reused for an NFS version
field.

Adapted from changes proposed by OAM7575 in PR #408.

Issue: #408, #354

Co-authored-by: OAM7575 <oam7575@gmail.com>

Co-authored-by: Shasheen Ediriweera <shasheene@gmail.com>