Improve failed to make backup dir message (4/5)
Reuses the existing (already translated) failed to write destination file
message when the first backup folder is made.

This is useful since this first step often causes a permission denied error,
especially on Linux environments, so it makes sense to have a clear message.

This commit doesn't have to be part of the 5 commit set, but moving it would
cause rebase conflicts.