integration-test: Fix EFI boot from DVD with hack
Implements a fix that I developed to workaround a bug in VirtualBox [1] where
EFI systems do not respect the boot order.

My workaround involves detaching the virtual hard drive (the sole hard drive of
the integration test), booting from DVD and reattaching the virtual hard drive.

VBoxManage complains that "The machine is not mutable (state is Running)", but
the solution works (at least with VirtualBox 7.0.12) and unblocks continued
integration-test development.

[1] https://www.virtualbox.org/ticket/19364