Fix missing second arg on large MBR gap, EFI NVRAM
Fixes the display_status() function call that's missing a second argument on
the code path for handling a large MBR gap, and the code path for updating the
EFI NVRAM during a restore.

The bug for both were latent in the codebase for Rescuezilla v2.4.2 and for
many earlier versions. It certainly will cause a runtime exception for the
display status function, however this was/is wrapped by the GLib idle_add()
function, which eats runtime exceptions without even a log statement. Therefore
this issue had no user impact, and was only broken for the (as of writing
unreleased) command-line interface after the recent minor refactor.