Install polkitd-pkla on Mantic so elevations works
Installs "polkitd-pkla" which is a "suggested" package for polkitd 123-1 on
Ubuntu 23.10 (Mantic) [1] [2], but was not present on Ubuntu 22.04 (Jammy) and
Ubuntu 23.04 (Lunar).

Without this package, the elevated GUI apps (Rescuezilla frontend, terminal,
rightclick -> pcmanfm) on the Ubuntu 23.10 (Mantic) build didn't work.

The description is listed as: Legacy "local authority" (.pkla) backend for
polkitd.

[1] https://packages.ubuntu.com/mantic/polkitd

[2] https://web.archive.org/web/20231125203746/https://packages.ubuntu.com/mantic/polkitd