Adds early build based on Ubuntu 24.04 LTS (Noble)
Adds a build based on the (yet-unreleased) Ubuntu 24.04 (Noble) for better
support of recent hardware.

The version specific packages installed based on the Ubuntu 23.10 (Mantic)
list.

Note: due to "A later version is already installed" error for repackaged Ubuntu
partclone v0.3.27, a slight hack was added to build.sh to delete the
Rescuezilla built deb file, which will be removed in future.