Stop ntfsfix during backup on source NTFS fs
No longer run an 'ntfsfix' command on NTFS partitions ON THE SOURCE DISK during
backup operations. The command continues to be run on the destination disk
after restoring though.

This is very likely the root cause of a critical severity error on a small
number of (hibernated) Windows disks where creating a backup image causes some
kind of corruption that can break the Windows boot with a Blue Screen of Death
-- possibly within the BCD data. A reported issue I have spent many hours
trying to reproduce but have yet been unsuccessful in reproduce.

Issue: #466

[1]
https://github.com/rescuezilla/rescuezilla/blob/2.5/src/apps/rescuezilla/rescuezilla/usr/lib/python3/dist-packages/rescuezilla/utility.py#L810