Stop bkg backup/clone if boot reserve check fails
Fixes a bug where if the Windows boot reserve function fails the wizard can
jump to the summary page (as intended) but the backup or clone operation
continues in the background leading to a disconnect in what the graphical
frontend is showing compared to underlying operation.

This is due to the main backup thread not returning in part because the
'completed_operation()' callback was occurring in a function call.

Fixes #501