Remove unnecessary gtk3-engines-breeze package
The package "gtk3-engines-breeze" is not available on Ubuntu 24.10 (Oracular).

However, looking at [1], the package was merely a transitional package that
contains nothing except representing a dependency on the 'breeze-gtk-theme'
package, which IS installed.

Therefore, it's safe to drop this transitional package as we still get the
package we actually care about.

[1] https://packages.ubuntu.com/jammy/gtk3-engines-breeze