backup: Validate 'lsblk --merge' with --help
Checks during backup whether lsblk supports "--merge" by running "--help" and
checking whether the "--merge" argument is present.

The "--merge" was introduced in [1] (included in util-linux v2.34, 2019-06-14)
introduced a --merge option to provide a fix to improve RAID listing [1].

This argument is not available in earlier versions of lsblk. And Ubuntu 18.04
(Bionic) release used for the Rescuezilla 32-bit release uses v2.31.1

Issue: #509

Issue: #531

[1] https://github.com/rescuezilla/rescuezilla/commit/5732eeb2ad622ba80598a4c9b35ebc7919fca935