Fix partclone-nbd executable building in ISO image
Attempt various fixes for why partclone-nbd build step is variously on GitHub
Actions (build container), local Linux environment but not in local Docker
container, apparently due to the "make install"  wanting to access
/usr/local/bin/partclone-nbd (a privileged location):

* Run make separately to the make install

* Explicitly do not use the Ubuntu release's build directory but build from the
folder above.

Note: the variables for the release and arch aren't set, and between local
Dockerfile and on GitHub Actions, there seems to be different behavior, causing
either the build to fail during the checkinstall step (due to path not
available) or the build to silently not include the partclone-nbd artifact.