Add Tamil language pack
Adds relevant language packages for the Tamil language (ta_IN) after a user
completed a translation of Rescuezilla into the Tamil language on Weblate
yesterday.

Chose 'us' keyboard as boot keyboard because could not find a custom Tamil
keyboard code, can update this in a future commit if there is a better choice.

Updates CHANGELOG with an additional note about a significant update in the
Norwegian translation.