Fix various (largely non user-facing) typos
Applies luzpaz's codespell tool to fix various user-facing and non-user-facing
typos.

The typos that are "user-facing" are in print statements that end up eg, in
`clonezilla-img` log file, so aren't particularly important regardless.

Found via [1].

I (@shasheene) have evaluated this codespell tool and deemed it not yet ready
to be included in Rescuezilla's build scripts, largely due to poor git
integration. (more detail on the PR #558).

[1] codespell -q 3 -S "*.po,*.desktop,CHANGELOG,./CHANGELOG.md" -L preprepared`