Bump partclone-nbd SHA, switch to CPack packaging
Bumps "partclone-nbd" git submodule to include the introduction of CPack to
create a DEB package.

Then switches to using a CPack-based build rather than 'checkinstall' which
wrapped the 'make install' for partclone-nbd.

This is being done now because the binary file isn't being included in the DEB
package anymore after the Docker container bump that coincided with Rescuezilla
v2.6.0's release.

It appears potentially a bug in 'checkinstall' that's causing it, so let's
switch to proper packaging with statically-defined configuration files rather
than checkinstall's (fantastic and very useful) more hackier packaging
approach.

Issue: #557