Fix GPG package for end-of-life Bionic and Focal
Manually specify the paths to the GPG public key needed for package
verification of Ubuntu 18.04 (Bionic) and Ubuntu 20.04 (Focal), in order to fix
the GPG package verification of these (end-of-life) releases after the keyring
no longer specifies the paths.

The failed verification issue was already observed in Ubuntu 18.04 (Bionic) in
December 2024, so signature check was temporarily disabled (#538). The same
issue has now occurred on the Focal build.

Further investigation revealed the manual path for package verification was all
that was needed to resolve this.

It appears to be caused by a combination of the debootstrap package being
updated and the build Docker container's base OS itself being recently updated
(the ubuntu-keyring package being updated).

Issue: #538