Add translation stub for SMB/CIFS version select
Finally adds some translation stubs for selecting the SMB/CIFS version.

Without this translators had no choice and the strings were presented in the
original English.

Also modified the string concatenation of the 'default' option in the source
code, so translators will translate it will full context instead of the
previous approach where it was reusing the string from the default compression
option.

Added with help of script [1]

[1] src/scripts$ ./insert-into-translations.py 'msgid "Remote path"'
newtranslate