Add sleep + sync to umount for 'target is busy'
Run a filesystem sync and with a (suboptimal) arbitrary 0.5 second sleep before
and after it.

This provides a hacky but functional workaround to the "umount:
/tmp/rescuezilla.ntfs/mount: target is busy" that can occur during backup
operations and cloning.

It happens consistently in the Rescuezilla Integration Test suite, but less
consistently in the live environment.

This doesn't fix the avoidance of the underlying eg, race condition because the
arbitrary 0.5 second sleep may not be enough on some systems. The change is
unfortunate and slow down eg, the shutdown of the app, but is sufficient for
now.

Issue: #561