Add new Javascript polkit rules file for Plucky
Introduces the new Javascript-based polkit rules file. This is required so that
on the upcoming Ubuntu 25.04 (Plucky) release the graphical desktop shortcuts
keep working that than silently failing to open.

The issue appears to be due to the removal of the deprecated 'polkit-pkla'
package that Rescuezilla has been relying on since Ubuntu 23.10 (Mantic) to
provide support for existing old-style approach.

Added a FIXME to remove 'polkit-pkla' from Ubuntu 24.04 (Noble) and Ubuntu
24.10 (Oracular), but removing it now for older platforms does not appear worth
the extra testing effort.