it: Check vboxsf mount, shared folder before mount
Guard to an observed issue in the integration test when the Clonezilla backup
operation is run after the Rescuezilla test where the second operation freezes
the process indefinitely, potentially in the interaction with the 'vboxsf'
kernel module.

Let's double-check the existence of the shared folder before the mkdir and
mount.