it: Create int test backup in one call, not two
Combine the integration test bash script's "remove previous backup" code with
the code that creates the backup image, and do this for both the create with
Rescuezilla and create with Clonezilla functions.

The benefit is this enables subsequent commands to occur within the Rescuezilla
Integration Test environment without a separate SSH invocation into the
integration test image.

Also tweaked the rm, to use sudo + safely delete even if variable unbound.