it: Bump timeout on Ubuntu.Windows.Dualboot.MBR
It was observed a workflow run [1] that Rescuezilla image that's successfully
created fails to restore with both Rescuezilla CLI and Clonezilla CLI due to
hitting a timeout error.

The timeouts on the self-hosted GitHub Actions runner were recently tightened,
and (unlike the Clonezilla backup case) there's not expected to be a separate
underlying issue here.

So the fix is to increase timeouts.

[1] https://github.com/rescuezilla/rescuezilla/actions/runs/16311778587/job/46073544570

Link will be dead due to log rotation soon, so here is the summary:

* Create backup with Rescuezilla: 23min (success)

* Create backup with Clonezilla: hit 90 minute timeout limit (failure, probably
  caused by a pipeline issue internally)

* Restore Rescuezilla with Rescuezilla CLI: hit 40 minute timeout failure

* Restore Rescuezilla with Clonezilla CLI: hit 60 minute timeout failure

* Restore Clonezilla with Rescuezilla CLI: N/A (due to failed backup above)

* Restore Clonezilla with Clonezilla CLI: N/A (due to failed backup above)