us: Support custom target branch when tagging
Enable the user to specify the branch name when running the create release
script, as the previous use of 'origin/master' is fine for an actual release,
but in the lead up to a release it's handy to specify a branch instead before
pushing to a forked remote.

This helps ensure everything still passes GitHub Actions continuous integration
before pushing publicly, especially around tagged versions which have some
somewhat complex interactions in the build system and GitHub Actions.