Reorder imports related to 'gi', ignore lint
There's a lint E402 (below) that would prefer the GObject Introspection import
be at the top of the file.

Unfortunately the need to run a 'require_version()' call makes this difficult.
Instead suboptimally, place this import after all the other imports (contrary
to the lint) and ignore the lint with a 'noqa' comment. This is against the
spirit of the lint, but it doesn't seem possible to place it at the top without
'noqa' all the other imports which then have a lint error so this seems fine.

$ ruff rule E402

Derived from the **pycodestyle** linter.

Checks for imports that are not at the top of the file.

According to [PEP 8], "imports are always put at the top of the file, just after any
module comments and docstrings, and before module globals and constants."

This rule makes an exception for both `sys.path` modifications (allowing for
`sys.path.insert`, `sys.path.append`, etc.) and `os.environ` modifications
between imports.

```python
"One string"
"Two string"
a = 1
import os
from sys import x
```

Use instead:
```python
import os
from sys import x

"One string"
"Two string"
a = 1
```

For Jupyter notebooks, this rule checks for imports that are not at the top of a *cell*.

[PEP 8]: https://peps.python.org/pep-0008/#imports