Test CombinedDriveState with snapshot + script
Updates the CombinedDriveState unit test (which had a commented out assertion)
with a snapshot testing style data collection from the integration-test VM
environment.

The reason the assertion was commented out was the parser for the lsblk output
was changed to add "SERIAL" as a column, but the unit test was never updated.

Since Clonezilla's "CombinedDriveState" concept combines data from many
different sources (eg, parted and sfdisk output for every disk), adds a bash
script to assist with collection of this snapshot-testing style data. The bash
script is intended to run manually in an integration test virtual machine
environment to assist in gathering input data. The script actually outputs on
stdout a long Python snippet and was copied in. Perhaps not the most elegant
approach, but should provide a close to zero effort way to keep things updated.